Replace wxTreeItemAttr and wxListItemAttr with wxItemAttr
The two existing structs were completely identical, just replace them with a
single wxItemAttr.

Notice that wxDataViewItemAttr is not quite the same, although pretty similar,
so it remains separate for now. It would be nice to combine it with this one
too in the future, e.g. to make it simpler to make items bold in a wxListCtrl.

# Conflicts:
#	Makefile.in
#	build/bakefiles/files.bkl
#	build/files
#	build/msw/wx_vc11_core.vcxproj
#	build/msw/wx_vc12_core.vcxproj.filters
#	include/wx/listbase.h
#	include/wx/qt/listctrl.h
#	samples/listctrl/listtest.h