Use custom draw in wxMSW wxHeaderCtrl to support colours
Add a helper wxMSWImpl::CustomDraw class which will be reused in the other
places too and, for now, use it just to implement support for custom colours
in wxHeaderCtrl.

Notice that the control took care of the custom font on its anyhow and that
background colour is ignored when themes are enabled, so the net effect of
this change is that now changing the header foreground colour works, while
it was ignored before.

# Conflicts:
#	Makefile.in
#	build/bakefiles/files.bkl
#	build/files
#	build/msw/makefile.bcc
#	build/msw/makefile.gcc
#	build/msw/makefile.vc
#	build/msw/wx_vc11_core.vcxproj
#	build/msw/wx_vc12_core.vcxproj.filters