wxIsPlatform64Bit: don't launch uname needlessly
Don't execute uname in Unix version of  wxIsPlatform64Bit () to
determine if the current platform is 64-bit capable if the binary is
already a 64bit one.

This is consistent with how MSW implementation behaves and avoids a
pointless invocation when running 64-bit binaries.

As an added benefit, this prevents user harassment by macOS 10.15 if
they launch a wx application from a "protected" location like ~/Desktop
or ~/Downloads - apparently stat()ing CWD is considered evil and
privacy-invading these days.