Don't set NSView.clipsToBounds = YES globally
In 991a74c wxWidgetImpl was changed to set clipsToBounds = YES globally,
to prevent new behavior in macOS 14, where views' rendering can extend
outside the view area, from applying to wxWidgets. This was to prevent
rendering artifacts (see e.g. #23916) because wxWindow rendering being
clipped to its area is arguably part of wxWindow API contract.

This however breaks rendering of several controls (most prominently
single-line wxTextCtrl) on macOS 26 Tahoe, and possibly more subtly
earlier too.

To avoid this, change the code to force clipsToBounds only on non-native
wxWindows and leave it as-is (which is false on macOS >= 14) for native
ones, so that they're rendering isn't disrupted.

Also add a semi-private wxWindow::MacClipsToBounds() method to allow
customizing this behavior. This is important for e.g. generic composite
controls that may be packed too tight and clip their children.

This fixes broken rendering of wxTextCtrl and wxSpinButton on macOS 26.